feat: Add Sales Report view
Introduces a new "Sales Report" view to track customer payments and billing information. This includes adding a new route to the application, updating the sidebar navigation, and defining new types for sale records. The Users component is updated to pass payment information to a new `addSale` function, which will be used to populate the sales report.

Also includes CSS for print visibility and minor adjustments to constants.